% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_calculate_champ.R
\name{calculate_champ}
\alias{calculate_champ}
\title{calculate_champ: Calculate the risk for 30 mortality using CHAMP}
\usage{
calculate_champ(
  rr,
  pulse,
  spo2,
  gcs,
  time_to_hems,
  age,
  cardiac_rhythm,
  medical_facility,
  vehicle_ground_unit,
  sex_male,
  code,
  limit_values = TRUE,
  errors_as_warnings = FALSE
)
}
\arguments{
\item{rr}{Systolic blood pressure (mmHg). Missing values allowed.}

\item{pulse}{Heart rate (bpm). Missing values allowed.}

\item{spo2}{Oxygen saturation (%) as digits e.g. 1% = 1 rather than 0.01.
Missing values allowed.}

\item{gcs}{Glasgow Coma Scale. Missing values allowed.}

\item{time_to_hems}{Time to HEMS arrival (minutes). Missing values allowed.}

\item{age}{Age in years.}

\item{cardiac_rhythm}{Cardiac rhythm (VF, VT, Asystole, PEA). Missing values allowed.}

\item{medical_facility}{Medical facility or nursing home as TRUE/FALSE or 1/0}

\item{vehicle_ground_unit}{HEMS vehicle ground unit (vs helicopter) as TRUE/FALSE or 1/0.}

\item{sex_male}{Patient sex male, TRUE/FALSE or 1/0.}

\item{code}{Patient group i.e. "cardiac arrest", "trauma", "respitory failure",
"chest pain", "stroke", "neurological", "psychiatric or intoxication", or "other". 
All other non-missing values are assumed to not be in any of those groups 
and instead belong to "Gynaecology and obstetrics" and "Infection" groups.}

\item{limit_values}{Should values be winsorized to match the original data set better? 
Default FALSE. Limits the values below / larger than 0.5th and 99.5th percentiles
to the percentile values. Note: Some variables such as *time_to_hems* can have
really long tales in outlier cases leading to questionable risk estimates.}

\item{errors_as_warnings}{Should error's be returned as warnings? Default FALSE.}
}
\value{
Vector of risks for each observation.
}
\description{
Calculates the CHAMP 30 day mortality risk for the given values.
All the variables need to be of the same length. NA values are allowed for 
*gcs*, *pulse*, *rr*, *cardiac_rhythm*, *spo2*, and *time_to_hems* as the 
calculator will automatically select the model fitted for those variables.
}
\examples{

calculate_champ(rr = 100,
               pulse = 100,
               spo2 = 100,
               gcs = 15,
               time_to_hems = 100,
               cardiac_rhythm = 1,
               age = 20,
               medical_facility = 0,
               vehicle_ground_unit = 1,
               sex_male = 1,
               code = "trauma")
calculate_champ(rr = 100,
               pulse = 100,
               spo2 = 100,
               gcs = 15,
               time_to_hems = 100,
               cardiac_rhythm = NA,
               age = 20,
               medical_facility = 0,
               vehicle_ground_unit = 1,
               sex_male = 1,
               code = "trauma")
calculate_champ(rr = rep(100, 3),
               pulse = rep(100, 3),
               spo2 = rep(100, 3),
               gcs = rep(15, 3),
               time_to_hems = rep(100, 3),
               cardiac_rhythm = rep(NA, 3),
               age = rep(20, 3),
               medical_facility = rep(0, 3),
               vehicle_ground_unit = rep(1, 3),
               sex_male = rep(1, 3),
               code = rep("trauma", 3))

calculate_champ(rr = c(100, 200, 300),
               pulse = rep(100, 3),
               spo2 = rep(100, 3),
               gcs = rep(15, 3),
               time_to_hems = rep(100, 3),
               cardiac_rhythm = rep(NA, 3),
               age = rep(20, 3),
               medical_facility = rep(0, 3),
               vehicle_ground_unit = rep(1, 3),
               sex_male = rep(1, 3),
               code = rep("trauma", 3),
               limit_values = TRUE)


}
