% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_calculate_champ_df.R
\name{calculate_champ_df}
\alias{calculate_champ_df}
\title{calculate_champ_df: Calculate the risk for 30 mortality using CHAMP with 
data frame as the input}
\usage{
calculate_champ_df(df_in)
}
\arguments{
\item{df_in}{Input data frame, needs to contain the following variables
"rr", "pulse", "code", "cardiac_rhythm", "spo2", 
"gcs", "time_to_hems", "age", 
"medical_facility", "vehicle_ground_unit", "sex_male"
and be in the correct format i.e. continuous variables as numeric, 
binary ones  as 0/1 and patient group code as string matching the defined 
categories.}
}
\value{
Vector of risks for each observation.
}
\description{
A fct function
}
\examples{

}
