% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coeffs}
\alias{coeffs}
\title{Coefficients for all the models}
\format{
A tibble with 64 rows and 1370 variables:
\describe{
  \item{set_name}{Name of the model set}
  \item{ind_X}{Indicator if variable X is used}
  \item{beta_X}{Regression coefficient for X variable}
  ...
}
}
\usage{
coeffs
}
\description{
Coefficients for the models used  in the CHAMP algorithm.
}
\keyword{datasets}
