% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var_names}
\alias{var_names}
\title{Names of the variables}
\format{
A tibble with 11 rows and 2 variables:
\describe{
  \item{data}{Name in the data}
  \item{func}{Name in the function}
}
}
\usage{
var_names
}
\description{
Names of the variables used in the CHAMP algorithm. Matches the variable 
definition excel to the functions.
}
\keyword{datasets}
