% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_calculate_champ.R
\name{calculate_champ}
\alias{calculate_champ}
\title{calculate_champ: Calculate the risk for 30 mortality using CHAMP}
\usage{
calculate_champ(df_in, df_coefs, return_risk = TRUE)
}
\arguments{
\item{df_in}{Input data frame}

\item{return_risk}{Should risk be returned? Default TRUE.}

\item{df_ind_coefs}{Data frame containing the coefficients for the models}
}
\value{
Vector of risks for each observation.
}
\description{
A fct function
}
\examples{


}
